creating a dictinct sections for users

+++ controllers/todosController.js
@@ -4,7 +4,7 @@ const todoSchema = require("../validator/todosValidator");
 // La liste de toutes les taches
 const getTodos = async (req, res) => {
   try {
-    const result = await pool.query("SELECT * FROM todos");
+    const result = await pool.query("SELECT * FROM todos WHERE user_id = $1", [req.user.id]);
     res.json(result.rows);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -20,8 +20,8 @@ const createTodo = async (req, res) => {
 
     const { tache } = req.body;
     const result = await pool.query(
-      "INSERT INTO todos (tache) VALUES ($1) RETURNING *",
-      [tache],
+      "INSERT INTO todos (tache, user_id) VALUES ($1, $2) RETURNING *",
+      [tache, req.user.id],
     );
     res.status(201).json(result.rows[0]);
   } catch (error) {
@@ -33,7 +33,7 @@ const createTodo = async (req, res) => {
 const deleteTodo = async (req, res) => {
   try {
     const id = parseInt(req.params.id);
-    await pool.query("DELETE FROM todos WHERE id = $1", [id]);
+    await pool.query("DELETE FROM todos WHERE id = $1 AND user_id = $2", [id, req.user.id]);
     res.json({ message: "Tache supprimée" });
   } catch (error) {
     res.status(500).json({ message: error.message });
